backend/requirements.txt
# FastAPI & server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-dotenv==1.1.0

# Data processing
pandas==2.2.3
numpy==2.2.1
openpyxl==3.1.5

# AI / ML
sentence-transformers==3.3.1
scikit-learn==1.6.1

# HTTP client
httpx==0.28.1

# Kaggle SDK
kaggle==1.6.17
